move sql functions to psql.c
