+++ pclsync/pbusinessaccount.c
@@ -31,6 +31,7 @@
 #include "plibs.h"
 #include "pnetlibs.h"
 #include "psys.h"
+#include "psql.h"
 
 #include <stdio.h>
