+++ pclsync/pfs.c
@@ -28,11 +28,14 @@
   DAMAGE.
 */
 
-#include <mbedtls/ctr_drbg.h>
-#include <mbedtls/debug.h>
-#include <mbedtls/entropy.h>
-#include <mbedtls/pkcs5.h>
-#include <mbedtls/ssl.h>
+#include <errno.h>
+#include <fuse.h>
+#include <pthread.h>
+#include <stdint.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <time.h>
 #include <pthread.h>
 #include <stdatomic.h>
 
@@ -59,15 +62,8 @@
 #include "psys.h"
 #include "ppath.h"
 #include "ptimer.h"
+#include "psql.h"
 
-#include <errno.h>
-#include <fuse.h>
-#include <pthread.h>
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <time.h>
 
 #ifndef FUSE_STAT
 #define FUSE_STAT stat
