+++ pclsync/pp2p.c
@@ -31,6 +31,7 @@
 
 #include <errno.h>
 #include <pthread.h>
+#include <string.h>
 
 #include "papi.h"
 #include "pfile.h"
@@ -47,7 +48,7 @@
 #include "psys.h"
 #include "ptimer.h"
 #include "putil.h"
-#include <string.h>
+#include "psql.h"
 
 #define P2P_ENCTYPE_RSA_AES 0
