+++ pclsync/psettings.c
@@ -29,12 +29,9 @@
    DAMAGE.
 */
 
-#include <mbedtls/ctr_drbg.h>
-#include <mbedtls/debug.h>
-#include <mbedtls/entropy.h>
-#include <mbedtls/pkcs5.h>
-#include <mbedtls/ssl.h>
 #include <pthread.h>
+#include <ctype.h>
+#include <string.h>
 
 #include "pfs.h"
 #include "plibs.h"
@@ -44,8 +41,7 @@
 #include "psettings.h"
 #include "ppath.h"
 #include "ptimer.h"
-#include <ctype.h>
-#include <string.h>
+#include "psql.h"
 
 typedef void (*setting_callback)();
 typedef void (*filter_callback)(void *);
