+++ pclsync/ptask.c
@@ -32,12 +32,6 @@
 #include <errno.h>
 #include <pthread.h>
 #include <stddef.h>
-
-#include <mbedtls/ctr_drbg.h>
-#include <mbedtls/debug.h>
-#include <mbedtls/entropy.h>
-#include <mbedtls/pkcs5.h>
-#include <mbedtls/ssl.h>
 #include <pthread.h>
 
 #include "papi.h"
@@ -55,6 +49,7 @@
 #include "ptask.h"
 #include "ptree.h"
 #include "pupload.h"
+#include "psql.h"
 
 #define get_len(t) (sizeof(t) - offsetof(t, request))
