+++ pclsync/pupload.c
@@ -29,11 +29,6 @@
    DAMAGE.
 */
 
-#include <mbedtls/ctr_drbg.h>
-#include <mbedtls/debug.h>
-#include <mbedtls/entropy.h>
-#include <mbedtls/pkcs5.h>
-#include <mbedtls/ssl.h>
 #include <pthread.h>
 #include <stddef.h>
 
@@ -57,6 +52,7 @@
 #include "ptools.h"
 #include "pupload.h"
 #include "putil.h"
+#include "psql.h"
 
 extern const unsigned char pfile_invalid_chars[];
